Added Buttons and Blank Space Scalability
GRect buttons were added and positioned on the left side of the screen. The hangpost was moved to the right a few hundred pixels to make room for the buttons. The blank space's formula's for spacing and scalability has been modified. Labels and functionality for the buttons have not yet been added or initialized.